requirement.txt
tensorflow-cpu
tensorflow_datasets
numpy
matplotlib
PyQt6
scikit-learn